File reading
Now lingwi can be fed with a filepath. When processing arguments, it'll be deduced if the input is a filename or a string. What's more, final output will be printed with the custom-made 'print' function; it's pretty bare-bones, for now supporting only cases where there's a newline character, and it doesn't account for more special whitespace characters.